refactor : 필요없는 상수 제거 및 패키지 이름 변경 utill -> generator
