+++ src/main/java/lotto/generator/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.generator;
 
 import java.util.List;
